Changed paths so their "pixel" positions would be whole. This way i can check for overlap and destroy any overlapping filler cells already on the path.  This way I can remove them without creating strange 0.5 gaps in the future. Small filler cells will disappear without leaving gaps.

+++ DungeonGenerator/Dungeon.cpp
@@ -52,7 +52,7 @@ void Dungeon::GenerateDungeon()
 	//constructCorridors();
 	{
 		srand((unsigned int)time(NULL));
-		float dx, dy, x, y;
+		int dx, dy, x, y;
 		Room *a, *b;
 		for (auto& outer : vTrueRooms) {
 			std::vector<Room*>& Edges = graph[outer];
@@ -93,7 +93,7 @@ void Dungeon::GenerateDungeon()
 	//Expand corridor sizes
 	{
 		for (auto room : vCorridorRooms)
-			room->expand(1);
+			 room->expand(1);
 	}
 	markAllTileMap(vCorridorRooms);
 }
